refactor(storage): update zettel tags and links without deletion operation
Previously, updating a zettel's tags and links took the crude approach
of deleting the existing links for the zettel, and then inserting the
most up to date tags and links for the zettel. These two updates now
take place by adding new links/tags and removing non-existent links/tags.

There is no need to explicitly perform an sql update on links and tags.
This is because a modified zettel link/tag will be result in both a add
and remove operation.