fix(storage): exclude directories with no readme files from the database
Consider a directory with a single pdf file. When its first procsessed,
only the directory will be inserted into the database, not the pdf file.
When processed for a second time, the zettel map is initialized by only
the zettel table; the dir table is not looked at. So this directory
would show up as not existing in the database and attempt to insert the
directory, creating a conflict on directory name.

Thus, the function meant to add a brand new zettel directory and all its
files now returns early when the given directory does not contain any
readme files.